Remove All use Unit instead

+++ modules/ingestor/src/test/scala/ClITest.scala
@@ -17,7 +17,7 @@ object CLITest extends weaver.FunSuite:
   test("index command"):
     expect(
       testCommand("index", "--index", "team", "--since", "0", "--until", "1", "--dry") ==
-        IndexOpts(Single(Index.Team), Instant.ofEpochSecond(0), Instant.ofEpochSecond(1), true).asRight
+        IndexOpts(Index.Team, Instant.ofEpochSecond(0), Instant.ofEpochSecond(1), true).asRight
     )
 
   test("watch command"):
